refactor all extraction into support.extract; extract also on SearchAllRecentlyAddedMissing
