+++ Contents/Code/support/plex_media.py
@@ -425,3 +425,4 @@ class PMSMediaProxy(object):
 
             m = m.children[0]
         return parts
+
